feat: Add ContainerTests for MySQLContainerConfig
